requirements.txt
# Core Framework & Utilities
fastapi
uvicorn
pydantic
python-dotenv
requests
click
colorama
typing-extensions
python-multipart
python-dateutil
six
tqdm
bcrypt

# AI & LLM
google-generativeai
langchain
langgraph
langchain-community
langchain-core
langchain-google-genai
langchain-openai
sentence-transformers
faiss-cpu
mem0ai
numpy

# Database & Storage
sqlalchemy
psycopg2-binary
pgvector
boto3
s3transfer

# Web Search & Scraping
duckduckgo-search
tavily-python
trafilatura
firecrawl-py
beautifulsoup4
bs4
lxml>=6
lxml_html_clean
ddgs

# PDF & Document Processing
pymupdf
pdf2image
easyocr
pypdf
PyPDF2
Pillow
pdf2docx
docx2pdf
python-docx
reportlab
pypandoc
wkhtmltopdf

# Google API Clients
google-api-python-client
google-auth
google-auth-httplib2
google-ai-generativelanguage
googleapis-common-protos

# Other Dependencies
cachetools
certifi
charset-normalizer
greenlet
grpcio
grpcio-status
h11
httplib2
httptools
paramiko
sshtunnel
idna
jmespath
proto-plus
protobuf
pyasn1
pyasn1_modules
pyparsing
PyYAML
rsa
soupsieve
starlette
uritemplate
urllib3
watchfiles
websockets
streamlit
altair
pydeck
watchdog
opencv-python-headless
